Finished switching Http client model. New API includes EbioticClient. All tests passing.
